requirements.txt
eth_account>=0.5.9
requests>=2.28.0
simplejson>=3.17.0
pyyaml>=6.0
pytest>=7.0.0

# Note: pysdk is included locally in the project (pysdk/ folder)
# The dependencies above are required by pysdk

